remove _extent_index.py
