+++ datacube/db_extent/__init__.py
@@ -1,1 +1,3 @@
-from ._extent_index import ExtentIndex
+from ._extent_utils import ExtentMetadata, compute_uuid, parse_time
+from ._extent_upload import ExtentUpload
+from ._extent_retrieve import ExtentIndex
